Update csv exporter
